64 bit tree

+++ src/Sir.Store/AnalyzedString.cs
@@ -8,7 +8,7 @@ namespace Sir
         public char[] Source { get; set; }
         public string Original { get; set; }
 
-        public IEnumerable<SortedList<int, byte>> Embeddings
+        public IEnumerable<SortedList<long, byte>> Embeddings
         {
             get
             {
